+++ packages/build-info/src/node/get-build-info.test.ts
@@ -17,13 +17,14 @@ test('should detect nothing in an empty project', async () => {
       "buildSystems": [],
       "frameworks": [],
       "jsWorkspaces": null,
+      "langRuntimes": [],
       "packageManager": null,
       "settings": [],
     }
   `)
 })
 
-test('should detect nothing in a simple golang project', async () => {
+test('should only detect go in a simple golang project', async () => {
   const cwd = mockFileSystem({
     'go.mod': '',
     'main.go': '',
@@ -34,6 +35,12 @@ test('should detect nothing in a simple golang project', async () => {
       "buildSystems": [],
       "frameworks": [],
       "jsWorkspaces": null,
+      "langRuntimes": [
+        {
+          "id": "go",
+          "name": "Go",
+        },
+      ],
       "packageManager": null,
       "settings": [],
     }
@@ -43,7 +50,14 @@ test('should detect nothing in a simple golang project', async () => {
 test('should not crash on invalid projects', async (ctx) => {
   const fixture = await createFixture('invalid-project', ctx)
   const { frameworks, packageManager } = await getBuildInfo({ projectDir: fixture.cwd })
-  expect(packageManager).toMatchInlineSnapshot
+  expect(packageManager).toMatchInlineSnapshot(`
+    {
+      "installCommand": "npm install",
+      "lockFile": "package-lock.json",
+      "name": "npm",
+      "runCommand": "npm run",
+    }
+  `)
   expect(frameworks).toEqual([])
 })
 
@@ -78,7 +92,7 @@ test.skipIf(platform() === 'win32')(
 )
 
 describe('Golang', () => {
-  test('should not detect anything inside a golang workspace', async (ctx) => {
+  test('should only detect Go inside a golang workspace', async (ctx) => {
     const fixture = await createFixture('go-workspace', ctx)
     const info = await getBuildInfo({ projectDir: 'bar', rootDir: fixture.cwd })
     expect(info).toMatchInlineSnapshot(`
@@ -86,6 +100,12 @@ describe('Golang', () => {
         "buildSystems": [],
         "frameworks": [],
         "jsWorkspaces": null,
+        "langRuntimes": [
+          {
+            "id": "go",
+            "name": "Go",
+          },
+        ],
         "packageManager": null,
         "settings": [],
       }