+++ packages/build-info/src/runtime/java.ts
@@ -1,0 +1,7 @@
+import { LangRuntime } from './runtime.js'
+
+export class Java extends LangRuntime {
+  id = 'java'
+  name = 'Java'
+  configFiles = ['project.clj', 'build.boot']
+}