+++ packages/build-info/src/runtime/node.ts
@@ -1,0 +1,7 @@
+import { LangRuntime } from './runtime.js'
+
+export class Node extends LangRuntime {
+  id = 'node'
+  name = 'NodeJS'
+  configFiles = ['.node-version', '.nvmrc', 'package.json']
+}