+++ packages/build-info/src/settings/get-build-settings.ts
@@ -15,16 +15,24 @@ export type Settings = {
     id: string
     name: string
   }
+  buildSystem?: {
+    id: string
+    name: string
+  }
   /** The dist directory that contains the build output */
   dist: string
   env: Record<string, string | undefined>
-  plugins: string[]
+  /** Plugins installed via the netlify.toml */
+  plugins_from_config_file: string[]
+  /** Plugins that are detected via the framework detection and therefore recommended */
+  plugins_recommended: string[]
   pollingStrategies: string[]
   /** The baseDirectory for the UI to configure (used to run the command in this working directory) */
   baseDirectory?: string
+  functionsDir?: string
   /** the workspace package path of an app */
   packagePath?: string
-  tomlModifications?: string
+  template?: string
 }
 
 async function applyBuildSystemOverrides(
@@ -40,6 +48,7 @@ async function applyBuildSystemOverrides(
     const build = cmds.find((cmd) => cmd.type === 'build')
     const dev = cmds.find((cmd) => cmd.type === 'dev')
     const dist = await buildSystem.getDist?.(baseDirectory)
+    const port = await buildSystem.getPort?.(baseDirectory)
 
     updatedSettings.name = `${buildSystem.name} + ${settings.name} ${baseDirectory}`
     if (build) {
@@ -54,6 +63,10 @@ async function applyBuildSystemOverrides(
       updatedSettings.dist = dist
     }
 
+    if (port !== undefined && port !== null) {
+      updatedSettings.frameworkPort = port
+    }
+
     // if the build system should be run from the root then set the base directory to an empty string
     // only applicable if we have a build or dev command for it
     if (buildSystem.runFromRoot && build && dev) {
@@ -64,9 +77,8 @@ async function applyBuildSystemOverrides(
   return updatedSettings
 }
 
-async function getSettings(framework: Framework, project: Project, baseDirectory: string): Promise<Settings> {
-  const frameworkDist = framework.staticAssetsDirectory || framework.build.directory
-
+/** Retrieve the settings for a framework, used inside the CLI in conjunction with getFramework */
+export async function getSettings(framework: Framework, project: Project, baseDirectory: string): Promise<Settings> {
   const devCommands = framework.getDevCommands()
   const buildCommands = framework.getBuildCommands()
 
@@ -75,9 +87,10 @@ async function getSettings(framework: Framework, project: Project, baseDirectory
     buildCommand: buildCommands[0],
     devCommand: devCommands[0],
     frameworkPort: framework.dev?.port,
-    dist: project.fs.join(baseDirectory, frameworkDist),
+    dist: project.fs.join(baseDirectory, framework.build.directory),
     env: framework.env || {},
-    plugins: framework.plugins || [],
+    plugins_from_config_file: [],
+    plugins_recommended: framework.plugins || [],
     framework: {
       id: framework.id,
       name: framework.name,
@@ -88,7 +101,7 @@ async function getSettings(framework: Framework, project: Project, baseDirectory
   }
 
   if (baseDirectory?.length && project.workspace?.isRoot) {
-    settings.dist = project.fs.join(baseDirectory, frameworkDist)
+    settings.dist = project.fs.join(baseDirectory, framework.build.directory)
   }
 
   // 1. try to apply overrides for package managers (like npm, pnpm or yarn workspaces)
@@ -102,6 +115,7 @@ async function getSettings(framework: Framework, project: Project, baseDirectory
 }
 
 export async function getBuildSettings(project: Project): Promise<Settings[]> {
+  project.logger.debug('[get-build-settings.ts] getBuildSettings')
   if (project.frameworks === undefined) {
     throw new Error('Please run the framework detection before calling the build settings!')
   }