+++ packages/build-info/src/workspaces/detect-workspace.ts
@@ -1,6 +1,7 @@
 import type { PackageJson } from 'read-pkg'
 import { parse } from 'yaml'
 
+import { FrameworkName } from '../frameworks/index.js'
 import { PkgManager } from '../package-managers/detect-package-manager.js'
 import { Project } from '../project.js'
 
@@ -9,6 +10,8 @@ import { getWorkspacePackages } from './get-workspace-packages.js'
 export type WorkspacePackage = {
   path: string
   name?: string
+  /** For some workspaces like Nx Integrated the build system is aware of the framework */
+  forcedFramework?: FrameworkName
 }
 
 export class WorkspaceInfo {