+++ packages/config/src/edge_functions.ts
@@ -5,14 +5,26 @@ const cacheValues = ['manual', 'off']
 export const validations = [
   {
     property: 'edge_functions.*',
-    ...validProperties(['path', 'function', 'cache'], []),
+    ...validProperties(['path', 'excludedPath', 'pattern', 'excludedPattern', 'function', 'cache'], []),
     example: () => ({ edge_functions: [{ path: '/hello', function: 'hello' }] }),
   },
   {
     property: 'edge_functions.*',
-    check: (edgeFunction) => edgeFunction.path !== undefined,
-    message: '"path" property is required.',
+    check: (edgeFunction) => edgeFunction.path !== undefined || edgeFunction.pattern !== undefined,
+    message: 'either "path" or "pattern" is required.',
     example: () => ({ edge_functions: [{ path: '/hello', function: 'hello' }] }),
+  },
+  {
+    property: 'edge_functions.*',
+    check: (edgeFunction) => !(edgeFunction.path !== undefined && edgeFunction.pattern !== undefined),
+    message: '"path" and "pattern" are mutually exclusive.',
+    example: () => ({ edge_functions: [{ path: '/hello', function: 'hello' }] }),
+  },
+  {
+    property: 'edge_functions.*',
+    check: (edgeFunction) => !(edgeFunction.excludedPath !== undefined && edgeFunction.excludedPattern !== undefined),
+    message: '"excludedPath" and "excludedPattern" are mutually exclusive.',
+    example: () => ({ edge_functions: [{ path: '/hello/*', function: 'hello', excludedPath: '/hello/no' }] }),
   },
   {
     property: 'edge_functions.*',
@@ -25,6 +37,28 @@ export const validations = [
     check: isString,
     message: 'must be a string.',
     example: () => ({ edge_functions: [{ path: '/hello', function: 'hello' }] }),
+  },
+  {
+    property: 'edge_functions.*.excludedPath',
+    check: (value) => isString(value) || (Array.isArray(value) && value.every(isString)),
+    message: 'must be a string or array of strings.',
+    example: () => ({
+      edge_functions: [{ path: '/products/*', excludedPath: ['/products/*.jpg'], function: 'customise' }],
+    }),
+  },
+  {
+    property: 'edge_functions.*.pattern',
+    check: isString,
+    message: 'must be a string.',
+    example: () => ({ edge_functions: [{ pattern: '/hello/(.*)', function: 'hello' }] }),
+  },
+  {
+    property: 'edge_functions.*.excludedPattern',
+    check: (value) => isString(value) || (Array.isArray(value) && value.every(isString)),
+    message: 'must be a string or array of strings.',
+    example: () => ({
+      edge_functions: [{ path: '/products/(.*)', excludedPattern: ['^/products/(.*)\\.jpg$'], function: 'customise' }],
+    }),
   },
   {
     property: 'edge_functions.*.function',