+++ packages/build-info/CHANGELOG.md
@@ -64,6 +64,124 @@
   * dependencies
     * @netlify/framework-info bumped from ^9.8.8 to ^9.8.9
 
+## [7.6.1](https://github.com/netlify/build/compare/build-info-v7.6.0...build-info-v7.6.1) (2023-07-19)
+
+
+### Bug Fixes
+
+* add correct 11ty logos for light and dark ([#5190](https://github.com/netlify/build/issues/5190)) ([c54f5c0](https://github.com/netlify/build/commit/c54f5c0556c317727f3529dd3effad2f0cc21329)), closes [#5187](https://github.com/netlify/build/issues/5187)
+
+## [7.6.0](https://github.com/netlify/build/compare/build-info-v7.5.0...build-info-v7.6.0) (2023-07-19)
+
+
+### Features
+
+* make error reporting configureable ([#5191](https://github.com/netlify/build/issues/5191)) ([5986af2](https://github.com/netlify/build/commit/5986af2f482e6d28ff342f56960f270d27af8c08))
+
+
+### Bug Fixes
+
+* fix nx-integrated detection via forced framework ([#5189](https://github.com/netlify/build/issues/5189)) ([ae6c09b](https://github.com/netlify/build/commit/ae6c09b1b889b48c71293d84d0e2422f003f5836))
+
+## [7.5.0](https://github.com/netlify/build/compare/build-info-v7.4.3...build-info-v7.5.0) (2023-07-19)
+
+
+### Features
+
+* update pnpm build command for workspaces to build dependents ([#5183](https://github.com/netlify/build/issues/5183)) ([aa9e363](https://github.com/netlify/build/commit/aa9e363e6a2d52c356e3f503cd2ee0a119aa03d8))
+
+## [7.4.3](https://github.com/netlify/build/compare/build-info-v7.4.2...build-info-v7.4.3) (2023-07-19)
+
+
+### Bug Fixes
+
+* detection of workspace file and correct dev port ([#5181](https://github.com/netlify/build/issues/5181)) ([7331c27](https://github.com/netlify/build/commit/7331c278861c633b8b99e315134a4d4b4e5846ae))
+
+## [7.4.2](https://github.com/netlify/build/compare/build-info-v7.4.1...build-info-v7.4.2) (2023-07-17)
+
+
+### Bug Fixes
+
+* cleanup feature flag for new framework detection as already rolled out to 100% ([#5169](https://github.com/netlify/build/issues/5169)) ([2858d70](https://github.com/netlify/build/commit/2858d70341a2632e41bb683d0ca200008ee6d0ba))
+
+## [7.4.1](https://github.com/netlify/build/compare/build-info-v7.4.0...build-info-v7.4.1) (2023-07-07)
+
+
+### Bug Fixes
+
+* downgrade dot-prop to support Node.js 14 ([#5147](https://github.com/netlify/build/issues/5147)) ([ff26375](https://github.com/netlify/build/commit/ff26375933e369c99add77c6f2f1a7e207a5c50a))
+* fix return type of get framework to return the class ([#5145](https://github.com/netlify/build/issues/5145)) ([8a1083e](https://github.com/netlify/build/commit/8a1083e08db98f345b1d7816c44458f9a702c3ce))
+
+## [7.4.0](https://github.com/netlify/build/compare/build-info-v7.3.5...build-info-v7.4.0) (2023-07-07)
+
+
+### Features
+
+* improve nx-integrated dist directory ([#5133](https://github.com/netlify/build/issues/5133)) ([3e99d21](https://github.com/netlify/build/commit/3e99d21e1369b7a2216746c241f6395ef226bdcc))
+
+## [7.3.5](https://github.com/netlify/build/compare/build-info-v7.3.4...build-info-v7.3.5) (2023-07-06)
+
+
+### Bug Fixes
+
+* export PollingStrategies and getSettings as part of the public api ([#5140](https://github.com/netlify/build/issues/5140)) ([0500193](https://github.com/netlify/build/commit/0500193686b715bf58cfb5204cdb1a754092c1bf))
+
+## [7.3.4](https://github.com/netlify/build/compare/build-info-v7.3.3...build-info-v7.3.4) (2023-07-06)
+
+
+### Bug Fixes
+
+* fixes an issue where package paths where not correctly resolved ([#5134](https://github.com/netlify/build/issues/5134)) ([48d9db2](https://github.com/netlify/build/commit/48d9db260a6f57f4a7dbd96504b369e2b3b2c9d2))
+
+## [7.3.3](https://github.com/netlify/build/compare/build-info-v7.3.2...build-info-v7.3.3) (2023-07-05)
+
+
+### Bug Fixes
+
+* fixes an issue with npm workspace filtering in the command ([#5125](https://github.com/netlify/build/issues/5125)) ([d161c0c](https://github.com/netlify/build/commit/d161c0c6d00a4be77864356c66a2f82af19c076a)), closes [#5123](https://github.com/netlify/build/issues/5123)
+
+## [7.3.2](https://github.com/netlify/build/compare/build-info-v7.3.1...build-info-v7.3.2) (2023-07-05)
+
+
+### Bug Fixes
+
+* fixes an issue where the wrong dist was used for frameworks with… ([#5122](https://github.com/netlify/build/issues/5122)) ([4533a61](https://github.com/netlify/build/commit/4533a61da4f40c302fedf8185ae8b85ac3e39fa6))
+
+## [7.3.1](https://github.com/netlify/build/compare/build-info-v7.3.0...build-info-v7.3.1) (2023-07-04)
+
+
+### Bug Fixes
+
+* place outside ff ([0898cf6](https://github.com/netlify/build/commit/0898cf6ceb6467fc1cbca3985f45457a720ffb24))
+
+## [7.3.0](https://github.com/netlify/build/compare/build-info-v7.2.0...build-info-v7.3.0) (2023-06-30)
+
+
+### Features
+
+* detect lang runtime ([#5105](https://github.com/netlify/build/issues/5105)) ([7242ddd](https://github.com/netlify/build/commit/7242dddef6307f1c673ba94f7026a85eb5a0fb59))
+
+## [7.2.0](https://github.com/netlify/build/compare/build-info-v7.1.1...build-info-v7.2.0) (2023-06-29)
+
+
+### Features
+
+* make events async ([#5106](https://github.com/netlify/build/issues/5106)) ([6fa5220](https://github.com/netlify/build/commit/6fa5220bc6d9f9e390564cd2f1a3a4c5bc9cd206))
+
+## [7.1.1](https://github.com/netlify/build/compare/build-info-v7.1.0...build-info-v7.1.1) (2023-06-29)
+
+
+### Bug Fixes
+
+* remove node specific dirname ([#5103](https://github.com/netlify/build/issues/5103)) ([bc5f991](https://github.com/netlify/build/commit/bc5f991955571f5e0b8a135cfa09ac1c01eff9c8))
+
+## [7.1.0](https://github.com/netlify/build/compare/build-info-v7.0.8...build-info-v7.1.0) (2023-06-29)
+
+
+### Features
+
+* **build-info:** add detection for toml files and settings collection ([#5102](https://github.com/netlify/build/issues/5102)) ([fcf221a](https://github.com/netlify/build/commit/fcf221a57da89b9b7888a929aa0f978ce716ed9d))
+
 ## [7.0.8](https://github.com/netlify/build/compare/build-info-v7.0.7...build-info-v7.0.8) (2023-06-15)